Selection events example appends to prev state, even w/ batched setState
Before this change, the eventLog appended the new event to the last
version of the state that updated the component. If multiple events are
fired from within the same event handler (e.g. a key press firing both
a start and end event), React batches setState calls rather than
updating the component with each call, meaning this.state does not
necessarily contain other updates from the same batch of calls. To
ensure events appended to the log aren't overwritten, we use the
prevState passed to the setState callback, which React guarantees to be
up to date.