Ignore mouseMove if no ongoing selection
Prior to this change, clicking & dragging within the window (but outside
the component) would update any existing selection. This change requires
that a click-and-drag must at least start within the component.

Note that selection actions (mouse or keyboard) *must* call both a
start/update method *and* endSelection to correctly track whether a
selection change is ongoing.