Only update DynamicData inputs when state changes
The NumberInputs in DynamicData are controlled, but also debounced - the state doesn't change until after 80ms of inactivity. When using value, cypress seems to 'type' too quickly, resulting in the wrong number being typed in. Using defaultValue fixes that, but that means the Reset button doesn't update the inputs, so we add the key prop to force it to render.