Simplify Examples and ExamplesPage routing
ExamplePage is now a simple component, rather than higher order
component. Because it takes props, we no longer use `component` Route
render method, but instead switch to using children. Because children
renders whether or not the Route matches, we also wrap all the example
Routes in a Switch (which ensures only one is picked).